requirements.txt
bidict==0.21.3
click==8.0.1
Flask==2.0.1
flask-marshmallow==0.14.0
Flask-SocketIO==5.1.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
marshmallow==3.13.0
marshmallow-sqlalchemy==0.26.1
PyJWT==2.1.0
python-engineio==4.2.1
python-socketio==5.4.0
six==1.16.0
SQLAlchemy==1.4.25
Werkzeug==2.0.1
